config: expose config struct
We may have to setup a few new headers inside config
