+++ compositor/config/config_lua.c
@@ -41,7 +41,7 @@
 #include "xdg.h"
 #include "lua_helper.h"
 #include "bindings.h"
-#include "config_internal.h"
+#include "config.h"
 
 static inline struct tw_config *
 to_user_config(lua_State *L)
